+++ lib/barcode_scanner_package.dart
@@ -1,5 +1,6 @@
 library;
 
 export 'src/barcode_scanner.dart';
+export 'src/simple_barcode_scanner.dart';
 export 'src/barcode_types.dart';
 export 'exceptions.dart';